+++ src/onegov.swissvotes/onegov/swissvotes/app.py
@@ -130,6 +130,9 @@ def get_frameworks_asset():
     yield 'react-dropdown-tree-select.css'
     yield 'form_dependencies.js'
     yield 'confirm.jsx'
+    yield 'jquery.datetimepicker.css'
+    yield 'jquery.datetimepicker.js'
+    yield 'datetimepicker.js'
 
 
 @SwissvotesApp.webasset('common')